+++ src/context/OrderContext.spec.tsx
@@ -2,9 +2,9 @@ import {act, fireEvent, render, screen, waitFor} from '@testing-library/react';
 import {jest} from '@jest/globals';
 import {createEmptyOrder, Order} from '../api/orders/Orders';
 import {useOrderContext, OrderProvider} from './OrderContext';
-import {OrderStatus, PaymentMethod} from '../api/orders/Orders';
 import type {OrderContextType} from '../@types/order.d';
 import {storage} from 'src/utils/localStorage';
+import {OrderStatus, PaymentMethod} from 'src/@types/order';
 
 function DummyComponent({futureOrder}: {futureOrder?: Order}) {
   const value = useOrderContext();
